Pass qrCode and targetCharacterId as string always

+++ app/src/main/java/org/shadowrunrussia2020/android/InteractWithBodyFragment.kt
@@ -15,9 +15,9 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.shadowrunrussia2020.android.character.CharacterViewModel
-import org.shadowrunrussia2020.android.common.utils.showErrorMessage
 import org.shadowrunrussia2020.android.common.models.FullQrData
 import org.shadowrunrussia2020.android.common.models.QrType
+import org.shadowrunrussia2020.android.common.utils.showErrorMessage
 import org.shadowrunrussia2020.android.model.qr.maybeQrScanned
 import org.shadowrunrussia2020.android.model.qr.startQrScan
 
@@ -65,7 +65,7 @@ class InteractWithBodyFragment : Fragment() {
                 mModel.postEvent(
                     "scanQr",
                     hashMapOf(
-                        "qrCode" to data.modelId.toInt(),
+                        "qrCode" to data.modelId,
                         "targetCharacterId" to args.targetId
                     )
                 )
